Crate multiboot2

Library that assists parsing the Multiboot2 Information Structure (MBI) from Multiboot2-compliant bootloaders, such as GRUB. It supports all tags from the specification including full support for the sections of ELF files. This library is no_std and can be used in a Multiboot2-kernel.

The GNU Multiboot(2) specification aims to provide a standardised method of sharing commonly used information about the host machine at boot time and give the payload, i.e. a kernel, a well defined machine state.

Example

use multiboot2::{BootInformation, BootInformationHeader};

fn kernel_entry(mb_magic: u32, mbi_ptr: u32) {
    if mb_magic == multiboot2::MAGIC {
        let boot_info = unsafe { BootInformation::load(mbi_ptr as *const BootInformationHeader).unwrap() };
        let _cmd = boot_info.command_line_tag();
    } else { /* Panic or use multiboot1 flow. */ }
}

MSRV

The MSRV is 1.70.0 stable.

Modules

builder

Module for the builder-feature.

Structs

BasicMemoryInfoTag

Basic memory info tag.

BootInformation

A Multiboot 2 Boot Information (MBI) accessor.

BootInformationHeader

The basic header of a boot information.

BootLoaderNameTag

The bootloader name tag.

CommandLineTag

This tag contains the command line string.

EFIBootServicesNotExitedTag

EFI ExitBootServices was not called tag.

EFIImageHandle32Tag

Tag that contains the pointer to the boot loader’s UEFI image handle (32-bit).

EFIImageHandle64Tag

Tag that contains the pointer to the boot loader’s UEFI image handle (64-bit).

EFIMemoryAreaType

The type of a memory range.

EFIMemoryAttribute

Flags describing the capabilities of a memory range.

EFIMemoryDesc

A structure describing a region of memory.

EFIMemoryMapTag

EFI memory map tag. The embedded EFIMemoryDescs follows the EFI specification.

EFISdt32Tag

EFI system table in 32 bit mode tag.

EFISdt64Tag

EFI system table in 64 bit mode tag.

ElfSection

A single generic ELF Section.

ElfSectionFlags

ELF Section bitflags.

ElfSectionIter

An iterator over some ELF sections.

ElfSectionsTag

This tag contains the section header table from an ELF binary.

EndTag

The end tag ends the information struct.

FramebufferColor

A framebuffer color descriptor in the palette.

FramebufferField

An RGB color type field.

FramebufferTag

The VBE Framebuffer information tag.

ImageLoadPhysAddrTag

The physical load address tag. Typically, this is only available if the binary was relocated, for example if the relocatable header tag was specified.

MemoryArea

A descriptor for an available or taken area of physical memory.

MemoryAreaTypeId

ABI-friendly version of MemoryAreaType.

MemoryMapTag

This tag provides an initial host memory map (legacy boot, not UEFI).

ModuleIter

An iterator over all module tags.

ModuleTag

The module tag can occur multiple times and specifies passed boot modules (blobs in memory). The tag itself doesn’t include the blog, but references its location.

RsdpV1Tag

This tag contains a copy of RSDP as defined per ACPI 1.0 specification.

RsdpV2Tag

This tag contains a copy of RSDP as defined per ACPI 2.0 or later specification.

SmbiosTag

This tag contains a copy of SMBIOS tables as well as their version.

Tag

Common base structure for all tags that can be passed via the Multiboot2 Information Structure (MBI) to a Multiboot2 payload/program/kernel.

TagTypeId

Serialized form of TagType that matches the binary representation (u32). The abstraction corresponds to the typ/type field of a Multiboot2 Tag. This type can easily be created from or converted to TagType.

VBECapabilities

The Capabilities field indicates the support of specific features in the graphics environment.

VBEControlInfo

VBE controller information.

VBEDirectColorAttributes

The DirectColorModeInfo field describes important characteristics of direct color modes.

VBEField

A VBE colour field.

VBEInfoTag

This tag contains VBE metadata, VBE controller information returned by the VBE Function 00h and VBE mode information returned by the VBE Function 01h.

VBEModeAttributes

A Mode attributes bitfield.

VBEModeInfo

Extended information about a specific VBE display mode from the mode list returned by VBEControlInfo (VBE Function 00h).

VBEWindowAttributes

The WindowAttributes describe the characteristics of the CPU windowing scheme such as whether the windows exist and are read/writeable, as follows:

Enums

ElfSectionType

An enum abstraction over raw ELF section types.

FramebufferType

The type of framebuffer.

MbiLoadError

Error type that describes errors while loading/parsing a multiboot2 information structure from a given address.

MemoryAreaType

Abstraction over defined memory types for the memory map as well as custom ones. Types 1 to 5 are defined in the Multiboot2 spec and correspond to the entry types of e820 memory maps.

StringError

Error type describing failures when parsing the string from a tag.

TagType

Higher level abstraction for TagTypeId that assigns each possible value to a specific semantic according to the specification. Additionally, it allows to use the TagType::Custom variant. It is not binary compatible with TagTypeId.

VBEMemoryModel

The MemoryModel field specifies the general type of memory organization used in modes.

Constants

MAGIC

Magic number that a Multiboot2-compliant boot loader will use to identify the handoff. The location depends on the architecture and the targeted machine state.

Traits

Pointee

Provides the pointer metadata type of any pointed-to type.

TagTrait

A trait to abstract over all sized and unsized tags (DSTs). For sized tags, this trait does not much. For DSTs, a TagTrait::dst_size implementation must be provided, which returns the right size hint for the dynamically sized portion of the struct.

Derive Macros

Pointee

Generates an implementation of Pointee for structs with a DST as its last field.