Enum multiboot2::MemoryAreaType

pub enum MemoryAreaType {
    Available,
    Reserved,
    AcpiAvailable,
    ReservedHibernate,
    Defective,
    Custom(u32),
}

Abstraction over defined memory types for the memory map as well as custom ones. Types 1 to 5 are defined in the Multiboot2 spec and correspond to the entry types of e820 memory maps.

This is not binary compatible with the Multiboot2 spec. Please use MemoryAreaTypeId instead.

Variants

Available

Available memory free to be used by the OS.

Reserved

A reserved area that must not be used.

AcpiAvailable

Usable memory holding ACPI information.

ReservedHibernate

Reserved memory which needs to be preserved on hibernation. Also called NVS in spec, which stands for “Non-Volatile Sleep/Storage”, which is part of ACPI specification.

Defective

Memory which is occupied by defective RAM modules.

Custom(u32)

Custom memory map type.