multiboot2

Struct ApmTag

pub struct ApmTag { /* private fields */ }

The Advanced Power Management (APM) tag.

Implementations

impl ApmTag

pub fn new( version: u16, cseg: u16, offset: u32, cset_16: u16, dset: u16, flags: u16, cseg_len: u16, cseg_16_len: u16, dseg_len: u16, ) -> Self

Creates a new tag.

pub const fn version(&self) -> u16

The version number of the APM BIOS.

pub const fn cseg(&self) -> u16

Contains the 16-bit code segment (CS) address for the APM entry point.

pub const fn offset(&self) -> u32

Represents the offset address within the code segment (cseg) for the APM entry point.

pub const fn cset_16(&self) -> u16

Contains the 16-bit code segment (CS) address used for 16-bit protected mode APM functions.

pub const fn dseg(&self) -> u16

Holds the 16-bit data segment (DS) address used by the APM BIOS for data operations.

pub const fn flags(&self) -> u16

Indicates the status and characteristics of the APM connection, such as if APM is present and its capabilities.

pub const fn cseg_len(&self) -> u16

Indicates the length, in bytes, of the data segment (dseg) used by the APM BIOS

pub const fn cseg_16_len(&self) -> u16

Provides the length, in bytes, of the 16-bit code segment (cseg_16) used for APM functions.

pub const fn dseg_len(&self) -> u16

Indicates the length, in bytes, of the data segment (dseg) used by the APM BIOS.