Struct multiboot2::BootLoaderNameTag

#[repr(C)]
pub struct BootLoaderNameTag { /* private fields */ }

The bootloader name tag.

Implementations

impl BootLoaderNameTag

pub fn new(name: &str) -> BoxedDst<Self>

pub fn name(&self) -> Result<&str, StringError>

Reads the name of the bootloader that is booting the kernel as Rust string slice without the null-byte.

For example, this returns "GRUB 2.02~beta3-5".

If the function returns Err then perhaps the memory is invalid.

Examples

if let Some(tag) = boot_info.boot_loader_name_tag() {
    assert_eq!(Ok("GRUB 2.02~beta3-5"), tag.name());
}