multiboot2

Struct BootdevTag

pub struct BootdevTag { /* private fields */ }

The end tag ends the information struct.

Implementations

impl BootdevTag

pub fn new(biosdev: u32, slice: u32, part: u32) -> Self

Creates a new tag.

pub const fn biosdev(&self) -> u32

Returns the bios device from which the device was booted from. 0x00 represents the first floppy disk. 0x80 represents the first hard disk, 0x81 the second hard disk, and so on.

pub const fn slice(&self) -> u32

The slice field identifies the partition (also known as a “slice” in BSD terminology) on the BIOS device from which the operating system was booted.

pub const fn part(&self) -> u32

The part field denotes the subpartition or logical partition within the primary partition (if applicable) from which the operating system was booted.