multiboot2

Struct DynSizedStructure

pub struct DynSizedStructure<H>
where
    H: Header,
{ /* private fields */ }

An C ABI-compatible dynamically sized type with a common sized Header and a dynamic amount of bytes without hidden implicit padding.

This structures combines a Header with the logically owned data by that header according to the reported Header::payload_len. Instances guarantees that the memory requirements promised in the crates description are respected.

This can be a Multiboot2 header tag, information tag, boot information, or a Multiboot2 header. It is the base for same-size casts to these corresponding structures using DynSizedStructure::cast. Depending on the context, the Header is different (header header, boot information header, header tag header, or boot information tag header).

ABI

This type uses the C ABI. The fixed Header portion is always there. Further, there is a variable amount of payload bytes. Thus, this type can only exist on the heap or references to it can be made by cast via fat pointers. The main constructor is DynSizedStructure::ref_from_bytes.

As terminating padding might be necessary for the proper Rust type layout, size_of_val(&self) might report additional padding bytes that are not reflected by the actual payload. These additional padding bytes however will be reflected in corresponding BytesRef instances from that this structure was created.

Implementations

impl<H> DynSizedStructure<H>
where H: Header,

pub fn ref_from_bytes( bytes: BytesRef<'_, H>, ) -> Result<&DynSizedStructure<H>, MemoryError>

Creates a new fat-pointer backed reference to a DynSizedStructure from the given BytesRef.

pub fn ref_from_slice( bytes: &[u8], ) -> Result<&DynSizedStructure<H>, MemoryError>

Creates a new fat-pointer backed reference to a DynSizedStructure from the given &[u8].

pub unsafe fn ref_from_ptr<'a>( ptr: NonNull<H>, ) -> Result<&'a DynSizedStructure<H>, MemoryError>

Creates a new fat-pointer backed reference to a DynSizedStructure from the given thin pointer to the Header. It reads the total size from the header.

Safety

The caller must ensure that the function operates on valid memory.

pub const fn header(&self) -> &H

Returns the underlying Header.

pub const fn payload(&self) -> &[u8]

Returns the underlying payload.

pub fn cast<T>(&self) -> &T
where T: MaybeDynSized<Header = H> + ?Sized,

Performs a memory-safe same-size cast from the base-structure to a specific MaybeDynSized. The idea here is to cast the generic mostly semantic-free version to a specific type with fields that have a clear semantic.

The provided T of type MaybeDynSized might be may be sized type or DST. This depends on the type. However, the source and the target both will have the same actual payload size and the same size_of_val.

Panic

Panics if base assumptions are violated. For example, the T of type MaybeDynSized must allow proper same-size casting to it.

Safety

This function is safe due to various sanity checks and the overall memory assertions done while constructing this type.

Panics

This panics if there is a size mismatch. However, this should never be the case if all types follow their documented requirements.