multiboot2

Struct EFIMemoryAttribute

pub struct EFIMemoryAttribute(/* private fields */);

Flags describing the capabilities of a memory range.

Implementations

impl MemoryAttribute

pub const UNCACHEABLE: MemoryAttribute

Supports marking as uncacheable.

pub const WRITE_COMBINE: MemoryAttribute

Supports write-combining.

pub const WRITE_THROUGH: MemoryAttribute

Supports write-through.

pub const WRITE_BACK: MemoryAttribute

Support write-back.

pub const UNCACHABLE_EXPORTED: MemoryAttribute

Supports marking as uncacheable, exported and supports the “fetch and add” semaphore mechanism.

pub const WRITE_PROTECT: MemoryAttribute

Supports write-protection.

pub const READ_PROTECT: MemoryAttribute

Supports read-protection.

pub const EXECUTE_PROTECT: MemoryAttribute

Supports disabling code execution.

pub const NON_VOLATILE: MemoryAttribute

Persistent memory.

pub const MORE_RELIABLE: MemoryAttribute

This memory region is more reliable than other memory.

pub const READ_ONLY: MemoryAttribute

This memory range can be set as read-only.

pub const SPECIAL_PURPOSE: MemoryAttribute

This memory is earmarked for specific purposes such as for specific device drivers or applications. This serves as a hint to the OS to avoid this memory for core OS data or code that cannot be relocated.

pub const CPU_CRYPTO: MemoryAttribute

This memory region is capable of being protected with the CPU’s memory cryptography capabilities.

pub const RUNTIME: MemoryAttribute

This memory must be mapped by the OS when a runtime service is called.

pub const ISA_VALID: MemoryAttribute

This memory region is described with additional ISA-specific memory attributes as specified in MemoryAttribute::ISA_MASK.

pub const ISA_MASK: MemoryAttribute

These bits are reserved for describing optional ISA-specific cache- ability attributes that are not covered by the standard UEFI Memory Attribute cacheability bits such as UNCACHEABLE, WRITE_COMBINE, WRITE_THROUGH, WRITE_BACK, and UNCACHEABLE_EXPORTED.

See Section 2.3 “Calling Conventions” in the UEFI Specification for further information on each ISA that takes advantage of this.

impl MemoryAttribute

pub const fn empty() -> MemoryAttribute

Get a flags value with all bits unset.

pub const fn all() -> MemoryAttribute

Get a flags value with all known bits set.

pub const fn bits(&self) -> u64

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u64) -> Option<MemoryAttribute>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u64) -> MemoryAttribute

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u64) -> MemoryAttribute

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<MemoryAttribute>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: MemoryAttribute) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: MemoryAttribute) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: MemoryAttribute)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: MemoryAttribute)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: MemoryAttribute)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: MemoryAttribute, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: MemoryAttribute) -> MemoryAttribute

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: MemoryAttribute) -> MemoryAttribute

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: MemoryAttribute) -> MemoryAttribute

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference( self, other: MemoryAttribute, ) -> MemoryAttribute

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> MemoryAttribute

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl MemoryAttribute

pub const fn iter(&self) -> Iter<MemoryAttribute>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<MemoryAttribute>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.