multiboot2

Struct ElfSectionsTag

pub struct ElfSectionsTag { /* private fields */ }

This tag contains the section header table from an ELF binary.

Implementations

impl ElfSectionsTag

pub fn new( number_of_sections: u32, entry_size: u32, shndx: u32, sections: &[u8], ) -> Box<Self>

Create a new ElfSectionsTag with the given data.

pub const fn sections(&self) -> ElfSectionIter<'_>

Get an iterator over the ELF sections.

pub const fn number_of_sections(&self) -> u32

Returns the amount of sections.

pub const fn entry_size(&self) -> u32

Returns the size of each entry.

pub const fn shndx(&self) -> u32

Returns the index of the section header string table.