Struct multiboot2::FramebufferTag

#[repr(C)]
pub struct FramebufferTag { /* private fields */ }

The VBE Framebuffer information tag.

Implementations

impl FramebufferTag

pub fn new( address: u64, pitch: u32, width: u32, height: u32, bpp: u8, buffer_type: FramebufferType<'_> ) -> BoxedDst<Self>

pub fn address(&self) -> u64

Contains framebuffer physical address.

This field is 64-bit wide but bootloader should set it under 4GiB if possible for compatibility with payloads which aren’t aware of PAE or amd64.

pub fn pitch(&self) -> u32

Contains the pitch in bytes.

pub fn width(&self) -> u32

Contains framebuffer width in pixels.

pub fn height(&self) -> u32

Contains framebuffer height in pixels.

pub fn bpp(&self) -> u8

Contains number of bits per pixel.

pub fn buffer_type(&self) -> Result<FramebufferType<'_>, UnknownFramebufferType>

The type of framebuffer, one of: Indexed, RGB or Text.