Struct multiboot2::MemoryMapTag

#[repr(C)]
pub struct MemoryMapTag { /* private fields */ }

This tag provides an initial host memory map (legacy boot, not UEFI).

The map provided is guaranteed to list all standard RAM that should be available for normal use. This type however includes the regions occupied by kernel, mbi, segments and modules. Kernel must take care not to overwrite these regions.

This tag may not be provided by some boot loaders on EFI platforms if EFI boot services are enabled and available for the loaded image (The EFI boot services tag may exist in the Multiboot2 boot information structure).

Implementations

impl MemoryMapTag

pub fn new(areas: &[MemoryArea]) -> BoxedDst<Self>

pub fn entry_size(&self) -> u32

Returns the entry size.

pub fn entry_version(&self) -> u32

Returns the entry version.

pub fn memory_areas(&self) -> &[MemoryArea]

Return the slice of the provided MemoryAreas.

Usually, this should already reflect the memory consumed by the code running this.