Struct multiboot2::ModuleTag

#[repr(C)]
pub struct ModuleTag { /* private fields */ }

The module tag can occur multiple times and specifies passed boot modules (blobs in memory). The tag itself doesn’t include the blog, but references its location.

Implementations

impl ModuleTag

pub fn new(start: u32, end: u32, cmdline: &str) -> BoxedDst<Self>

pub fn cmdline(&self) -> Result<&str, StringError>

Reads the command line of the boot module as Rust string slice without the null-byte. This is an null-terminated UTF-8 string. If this returns Err then perhaps the memory is invalid or the bootloader doesn’t follow the spec.

For example, this returns "--test cmdline-option".if the GRUB config contains "module2 /some_boot_module --test cmdline-option".

If the function returns Err then perhaps the memory is invalid.

pub fn start_address(&self) -> u32

Start address of the module.

pub fn end_address(&self) -> u32

End address of the module

pub fn module_size(&self) -> u32

The size of the module/the BLOB in memory.