multiboot2

Struct TagHeader

#[repr(C, align(8))]
pub struct TagHeader {
    pub typ: TagTypeId,
    pub size: u32,
}

The common header that all tags have in common. This type is ABI compatible.

Not to be confused with Multiboot header tags, which are something different.

It is the sized counterpart of GenericTag, an internal type.

Fields
typ: TagTypeId

The ABI-compatible TagType.

size: u32

The total size of the tag including the header.

Implementations

impl TagHeader

pub fn new(typ: impl Into<TagTypeId>, size: u32) -> Self

Creates a new header.