multiboot2

Struct VBEControlInfo

pub struct VBEControlInfo {
    pub signature: [u8; 4],
    pub version: u16,
    pub oem_string_ptr: u32,
    pub capabilities: VBECapabilities,
    pub mode_list_ptr: u32,
    pub total_memory: u16,
    pub oem_software_revision: u16,
    pub oem_vendor_name_ptr: u32,
    pub oem_product_name_ptr: u32,
    pub oem_product_revision_ptr: u32,
    /* private fields */
}

VBE controller information.

The capabilities of the display controller, the revision level of the VBE implementation, and vendor specific information to assist in supporting all display controllers in the field are listed here.

The purpose of this struct is to provide information to the kernel about the general capabilities of the installed VBE software and hardware.

Fields
signature: [u8; 4]

VBE Signature aka โVESAโ.

version: u16

The VBE version.

oem_string_ptr: u32

A far pointer the the OEM String.

capabilities: VBECapabilities

Capabilities of the graphics controller.

mode_list_ptr: u32

Far pointer to the video mode list.

total_memory: u16

Number of 64KiB memory blocks (Added for VBE 2.0+).

oem_software_revision: u16

VBE implementation software revision.

oem_vendor_name_ptr: u32

Far pointer to the vendor name string.

oem_product_name_ptr: u32

Far pointer to the product name string.

oem_product_revision_ptr: u32

Far pointer to the product revision string.