Enum multiboot2::ElfSectionType

#[repr(u32)]
pub enum ElfSectionType {

    Unused = 0,
    ProgramSection = 1,
    LinkerSymbolTable = 2,
    StringTable = 3,
    RelaRelocation = 4,
    SymbolHashTable = 5,
    DynamicLinkingTable = 6,
    Note = 7,
    Uninitialized = 8,
    RelRelocation = 9,
    Reserved = 10,
    DynamicLoaderSymbolTable = 11,
    EnvironmentSpecific = 1_610_612_736,
    ProcessorSpecific = 1_879_048_192,

}

An enum abstraction over raw ELF section types.

Variants

Unused = 0

This value marks the section header as inactive; it does not have an associated section. Other members of the section header have undefined values.

ProgramSection = 1

The section holds information defined by the program, whose format and meaning are determined solely by the program.

LinkerSymbolTable = 2

This section holds a linker symbol table.

StringTable = 3

The section holds a string table.

RelaRelocation = 4

The section holds relocation entries with explicit addends, such as type Elf32_Rela for the 32-bit class of object files. An object file may have multiple relocation sections.

SymbolHashTable = 5

The section holds a symbol hash table.

DynamicLinkingTable = 6

The section holds dynamic linking tables.

Note = 7

This section holds information that marks the file in some way.

Uninitialized = 8

A section of this type occupies no space in the file but otherwise resembles ProgramSection. Although this section contains no bytes, the sh_offset member contains the conceptual file offset.

RelRelocation = 9

The section holds relocation entries without explicit addends, such as type Elf32_Rel for the 32-bit class of object files. An object file may have multiple relocation sections.

Reserved = 10

This section type is reserved but has unspecified semantics.

DynamicLoaderSymbolTable = 11

This section holds a dynamic loader symbol table.

EnvironmentSpecific = 1_610_612_736

Values in this inclusive range ([0x6000_0000, 0x6FFF_FFFF)) are reserved for environment-specific semantics.

ProcessorSpecific = 1_879_048_192

Values in this inclusive range ([0x7000_0000, 0x7FFF_FFFF)) are reserved for processor-specific semantics.