multiboot2

Struct BootLoaderNameTag

pub struct BootLoaderNameTag { /* private fields */ }

The bootloader name tag.

Implementations

impl BootLoaderNameTag

pub fn new(name: &str) -> Box<Self>

Constructs a new tag.

pub fn typ(&self) -> TagType

Returns the underlying TagType.

pub const fn size(&self) -> usize

Returns the underlying tag size.

pub fn name(&self) -> Result<&str, StringError>

Reads the name of the bootloader that is booting the kernel as Rust string slice without the null-byte.

For example, this returns "GRUB 2.02~beta3-5".

If the function returns Err then perhaps the memory is invalid.

Examples

if let Some(tag) = boot_info.boot_loader_name_tag() {
    assert_eq!(Ok("GRUB 2.02~beta3-5"), tag.name());
}